cast cellids to integer

+++ scripts/ex-prt-mp7-p02.py
@@ -340,14 +340,14 @@ def build_gwf_sim():
 
     # Instantiate the MODFLOW 6 gwf well package
     welcells = ix.intersects(MultiPoint(wel_coords), dataframe=False)
-    welcells = welcells["cellids"]
+    welcells = welcells["cellids"].astype(int)
     welspd = [[(2, icpl), wel_q[idx]] for idx, icpl in enumerate(welcells)]
     flopy.mf6.ModflowGwfwel(gwf, print_input=True, stress_period_data=welspd)
 
     # Instantiate the MODFLOW 6 gwf river package
     riverline = [(Lx - 1.0, Ly), (Lx - 1.0, 0.0)]
     rivcells = ix.intersects(LineString(riverline), dataframe=False)
-    rivcells = rivcells["cellids"]
+    rivcells = rivcells["cellids"].astype(int)
     rivspd = [
         [(0, icpl), riv_h, riv_c, riv_z, riv_iface, riv_iflowface] for icpl in rivcells
     ]
